issues/24 - adding more tests
